refactor: close move `@inlang/core/fs` back to `@inlang/core/config` #471
