+++ source-code/core/package.json
@@ -14,10 +14,6 @@
 			"types": "./dist/config/index.d.ts",
 			"import": "./dist/config/index.js"
 		},
-		"./fs": {
-			"types": "./dist/fs/index.d.ts",
-			"import": "./dist/fs/index.js"
-		},
 		"./lint": {
 			"types": "./dist/lint/index.d.ts",
 			"import": "./dist/lint/index.js"
