--- source-code/core/src/config/environment-functions/index.ts
@@ -1,1 +1,0 @@
-export { initialize$import, type $import } from "./$import.js"
