+++ source-code/core/src/config/index.ts
@@ -1,2 +1,2 @@
-export * from "./schema.js"
-export * from "./environment-functions/index.js"
+export { Config, EnvironmentFunctions, DefineConfig } from "./schema.js"
+export { initialize$import } from "./environment-functions/$import.js"
